# backend/openapi_spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI小说创作系统",
    "description": "基于多Agent协作的智能小说创作平台",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/story-bible/facts": {
      "post": {
        "tags": [
          "Story Bible"
        ],
        "summary": "Create Fact",
        "description": "创建故事事实。",
        "operationId": "create_fact_api_story_bible_facts_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FactCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FactResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Story Bible"
        ],
        "summary": "List Facts",
        "description": "获取小说的事实列表,可按状态过滤。",
        "operationId": "list_facts_api_story_bible_facts_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "novel_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "Novel Id"
            }
          },
          {
            "name": "status_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(active|retired)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Filter"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FactResponse"
                  },
                  "title": "Response List Facts Api Story Bible Facts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/story-bible/facts/{fact_id}": {
      "get": {
        "tags": [
          "Story Bible"
        ],
        "summary": "Get Fact",
        "description": "获取单条事实。",
        "operationId": "get_fact_api_story_bible_facts__fact_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "fact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Fact Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FactResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Story Bible"
        ],
        "summary": "Update Fact",
        "description": "更新事实(支持 active/retired 状态流转)。",
        "operationId": "update_fact_api_story_bible_facts__fact_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "fact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Fact Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FactUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FactResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Story Bible"
        ],
        "summary": "Delete Fact",
        "description": "删除事实。",
        "operationId": "delete_fact_api_story_bible_facts__fact_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "fact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Fact Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/story-bible/events": {
      "post": {
        "tags": [
          "Story Bible"
        ],
        "summary": "Create Event",
        "description": "创建剧情事件。",
        "operationId": "create_event_api_story_bible_events_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Story Bible"
        ],
        "summary": "List Events",
        "description": "获取小说的事件列表,按故事天数排序。",
        "operationId": "list_events_api_story_bible_events_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "novel_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "Novel Id"
            }
          },
          {
            "name": "status_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(planned|occurred)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Filter"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventResponse"
                  },
                  "title": "Response List Events Api Story Bible Events Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/story-bible/events/{event_id}": {
      "get": {
        "tags": [
          "Story Bible"
        ],
        "summary": "Get Event",
        "description": "获取单个事件。",
        "operationId": "get_event_api_story_bible_events__event_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Story Bible"
        ],
        "summary": "Update Event",
        "description": "更新事件。",
        "operationId": "update_event_api_story_bible_events__event_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Story Bible"
        ],
        "summary": "Delete Event",
        "description": "删除事件。",
        "operationId": "delete_event_api_story_bible_events__event_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "EventCreate": {
        "properties": {
          "novel_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Novel Id",
            "description": "所属小说ID"
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "description": "事件标题"
          },
          "description": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 1,
            "title": "Description",
            "description": "事件描述"
          },
          "story_day": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Story Day",
            "description": "故事内天数",
            "default": 1
          },
          "chapter": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapter",
            "description": "对应章节号"
          },
          "involved_characters": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 50,
            "title": "Involved Characters",
            "description": "关联角色名列表"
          },
          "foreshadowing": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000
              },
              {
                "type": "null"
              }
            ],
            "title": "Foreshadowing",
            "description": "埋下的伏笔/线索"
          },
          "status": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(planned|occurred)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status",
            "description": "planned 或 occurred"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "novel_id",
          "title",
          "description"
        ],
        "title": "EventCreate",
        "description": "创建剧情事件。"
      },
      "EventResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "novel_id": {
            "type": "integer",
            "title": "Novel Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "story_day": {
            "type": "integer",
            "title": "Story Day"
          },
          "chapter": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapter"
          },
          "involved_characters": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Involved Characters",
            "default": []
          },
          "foreshadowing": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Foreshadowing"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "novel_id",
          "title",
          "description",
          "story_day",
          "status",
          "created_at"
        ],
        "title": "EventResponse",
        "description": "事件读取模型。"
      },
      "EventUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 8000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "story_day": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Story Day"
          },
          "chapter": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapter"
          },
          "involved_characters": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Involved Characters"
          },
          "foreshadowing": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000
              },
              {
                "type": "null"
              }
            ],
            "title": "Foreshadowing"
          },
          "status": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(planned|occurred)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "EventUpdate",
        "description": "更新剧情事件。"
      },
      "FactCreate": {
        "properties": {
          "novel_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Novel Id",
            "description": "所属小说ID"
          },
          "subject": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Subject",
            "description": "主体(角色/地点/组织名)"
          },
          "attribute": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Attribute",
            "description": "属性(如身份、位置、持有物)"
          },
          "value": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Value",
            "description": "事实值"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "补充说明"
          },
          "chapter_established": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapter Established",
            "description": "确立该事实的章节号"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "novel_id",
          "subject",
          "attribute",
          "value"
        ],
        "title": "FactCreate",
        "description": "创建故事事实:状态固定从 active 开始。"
      },
      "FactResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "novel_id": {
            "type": "integer",
            "title": "Novel Id"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "attribute": {
            "type": "string",
            "title": "Attribute"
          },
          "value": {
            "type": "string",
            "title": "Value"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "chapter_established": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapter Established"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "retired_chapter": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retired Chapter"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "novel_id",
          "subject",
          "attribute",
          "value",
          "status",
          "created_at"
        ],
        "title": "FactResponse",
        "description": "事实读取模型:不设写入预算,历史数据可完整返回。"
      },
      "FactUpdate": {
        "properties": {
          "value": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "chapter_established": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapter Established"
          },
          "status": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(active|retired)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status",
            "description": "active 或 retired"
          },
          "retired_chapter": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Retired Chapter",
            "description": "失效章节号"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "FactUpdate",
        "description": "更新事实:支持状态流转 active/retired。"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}